third_party/WebKit/LayoutTests/fast/multicol/span/invalid-span-1.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<div id="elm" style="-webkit-column-span:all; -webkit-column-span:1;"></div>
<script>
    description("column-span only takes the values 'none' and 'all'. '1' is invalid.");
    shouldBe("getComputedStyle(document.getElementById('elm')).WebkitColumnSpan", "'all'");
</script>
