third_party/blink/web_tests/platform/generic/fast/multicol/span/invalid-span-1-expected.txt

column-span only takes the values 'none' and 'all'. '1' is invalid.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.getElementById('elm')).WebkitColumnSpan is 'all'
PASS successfullyParsed is true

TEST COMPLETE
